a8n: remove namespace field, always use current user
Right now, the namespace field in the campaign form is not useful. Site admins always see all campaigns on the campaign list page, and non-site admins can't view any campaigns at all. It adds needless complexity to the form. This commit makes all campaigns created in the current user's namespace. When we support campaign namespaces throughout the UI, we can add this field back (or something like it).